# authMicro/docs/swagger.yaml
basePath: /auth/v1
definitions:
  entity.BaseError:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  entity.BaseValidationError:
    properties:
      code:
        type: string
      fieldErrors:
        items:
          $ref: '#/definitions/entity.FieldError'
        type: array
      name:
        type: string
    type: object
  entity.FieldError:
    properties:
      message:
        type: string
      nameField:
        type: string
    type: object
  github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.LoginResponse:
    properties:
      codeExpires:
        type: integer
      codePattern:
        type: string
    type: object
  github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.RegistrationResponse:
    properties:
      codeExpires:
        type: integer
      codePattern:
        type: string
    type: object
  github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.TokensResponse:
    properties:
      accessToken:
        type: string
      accessTokenExpires:
        type: integer
      refreshToken:
        type: string
      refreshTokenExpires:
        type: integer
    type: object
info:
  contact:
    name: Pavel Sheludyakov
    url: https://github.com/PavelShe11/studBridge
  description: 'Сервис аутентификации: регистрация, вход по OTP-коду, управление JWT
    токенами.'
  title: StudBridge Auth API
  version: "1.0"
paths:
  /login/confirmEmail:
    post:
      consumes:
      - application/json
      description: |-
        Проверяет OTP-код и возвращает пару JWT токенов (access + refresh).
        Обязательные поля: email (string), code (string).
      parameters:
      - description: Email и OTP-код
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.BaseValidationError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.BaseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.BaseError'
      summary: Подтвердить вход
      tags:
      - login
  /login/sendCodeEmail:
    post:
      consumes:
      - application/json
      description: |-
        Находит аккаунт по email и генерирует OTP-код для входа.
        Обязательное поле: email (string).
      parameters:
      - description: Email пользователя
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.BaseValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.BaseError'
      summary: Отправить код входа
      tags:
      - login
  /refreshToken:
    post:
      consumes:
      - application/json
      description: |-
        Проверяет refresh токен и возвращает новую пару JWT токенов (access + refresh). Старый refresh токен инвалидируется.
        Обязательное поле: refreshToken (string).
      parameters:
      - description: Refresh токен
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.TokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.BaseError'
        "401":
          description: Токен недействителен или истёк
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.BaseError'
      summary: Обновить токены
      tags:
      - tokens
  /registration:
    post:
      consumes:
      - application/json
      description: |-
        Создаёт или продлевает сессию регистрации и генерирует OTP-код.
        Набор полей в теле запроса определяется правилами валидации user-service (ValidateAccountData).
        Auth service передаёт данные напрямую — без фиксированной схемы.
      parameters:
      - description: Данные пользователя (схема определяется user-service)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PavelShe11_studbridge_authMicro_internal_infrastructure_inbound_rest_models.RegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.BaseValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.BaseError'
      summary: Отправить код регистрации
      tags:
      - registration
  /registration/confirmEmail:
    post:
      consumes:
      - application/json
      description: |-
        Проверяет OTP-код из сессии регистрации, затем передаёт данные в user-service для создания аккаунта.
        Набор полей определяется user-service. Обязательные поля: email (string), code (string).
      parameters:
      - description: Email, OTP-код и данные пользователя (схема определяется user-service)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Регистрация успешно завершена
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.BaseValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.BaseError'
      summary: Подтвердить email при регистрации
      tags:
      - registration
swagger: "2.0"
tags:
- description: Двухшаговая регистрация с OTP-кодом
  name: registration
- description: Двухшаговый вход с OTP-кодом
  name: login
- description: Управление JWT токенами
  name: tokens